Fixes a bug in health check with service pagination
This fixes an issue with ECS clusters that run more than 10
unique services, as the API only returns the first 10 services.

The previous version did not check the load balancer health for
all services running on the cluster, which can potentially cause
an outage if the cluster is cycled too quickly.

This also adds a --verbose flag to the health check which can be
useful for troubleshooting which service/load balancer is unhealthy.