Improves the service health check
This improves the service health check by checking the runningCount
against the desiredCount, which helps in the case where an application
is running at a lower scale.

This would report that the service is unhealthy and prevent the next
node from being terminated before the task is running on another ECS
instance.