Add retry logic while new ECS instances come online
The immediate exit logic on failure is too aggressive when
containers need to be shuffled on ECS instances. This is
especially noticeable on clusters where the services run
at the same scale as the cluster, so container instances
need to be moved to another ECS instance.

This fixes the case where a new ECS instance comes online
(which satisfies the desired count for the ECS cluster),
but the services are not healthy because containers are still
starting up or are still being registered in the load balancer.