Change the workflow interval to 60 seconds
Hitting the AWS APIs every 10 seconds is a bit excessive because
it takes much longer than that to provision a new EC2 instance
and for it to become available to the ECS cluster.

For a large ECS cluster, it also needs to check the load balancer
health for every service on the cluster, so this can make a lot
of API requests.

60 seconds is a bit more reasonable.