Check for failed tasks on new ECS instances
Cancel the cluster upgrade if there are failed tasks on the new
ECS instance.

The logic checks if the cluster scale is the desired scale and
if there are pending tasks, it checks to see if there are any
failed tasks on the new ECS instance.

This is less aggressive with immediately quitting the upgrade
process, but handles the case where containers can't start on the
new ECS instance and prevents users from pushing a bad server
configuration.